+++ src/libstd/cleanup.rs
@@ -10,23 +10,17 @@
 
 #[doc(hidden)];
 
-use libc::{c_char, c_void, intptr_t, uintptr_t};
-use ptr::mut_null;
+use libc::{c_char, intptr_t, uintptr_t};
+use ptr::{mut_null, to_unsafe_ptr};
 use repr::BoxRepr;
-use sys::TypeDesc;
 use cast::transmute;
 #[cfg(not(test))] use unstable::lang::clear_task_borrow_list;
-
-#[cfg(not(test))] use ptr::to_unsafe_ptr;
 
 /**
  * Runtime structures
  *
  * NB: These must match the representation in the C++ runtime.
  */
-
-type DropGlue<'self> = &'self fn(**TypeDesc, *c_void);
-type FreeGlue<'self> = &'self fn(**TypeDesc, *c_void);
 
 type TaskID = uintptr_t;
 
@@ -164,6 +158,20 @@ fn debug_mem() -> bool {
     false
 }
 
+#[cfg(stage0)]
+unsafe fn call_drop_glue(tydesc: *::std::unstable::intrinsics::TyDesc, data: *i8) {
+    use sys::TypeDesc;
+
+    let tydesc: *TypeDesc = transmute(tydesc);
+    let drop_glue: extern "Rust" fn(**TypeDesc, *i8) = transmute((*tydesc).drop_glue);
+    drop_glue(to_unsafe_ptr(&tydesc), data);
+}
+
+#[cfg(not(stage0))]
+unsafe fn call_drop_glue(tydesc: *::std::unstable::intrinsics::TyDesc, data: *i8) {
+    ((*tydesc).drop_glue)(to_unsafe_ptr(&tydesc), data);
+}
+
 /// Destroys all managed memory (i.e. @ boxes) held by the current task.
 #[cfg(not(test))]
 #[lang="annihilate"]
@@ -205,9 +213,7 @@ pub unsafe fn annihilate() {
     // callback, as the original value may have been freed.
     for each_live_alloc(false) |box, uniq| {
         if !uniq {
-            let tydesc: *TypeDesc = transmute(copy (*box).header.type_desc);
-            let drop_glue: DropGlue = transmute(((*tydesc).drop_glue, 0));
-            drop_glue(to_unsafe_ptr(&tydesc), transmute(&(*box).data));
+            call_drop_glue((*box).header.type_desc, transmute(&(*box).data));
         }
     }
